requirements.txt
Django>=4.2.0,<5.0
Pillow>=10.0.0
reportlab>=4.0.0
django-hcaptcha>=0.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
openpyxl>=3.1.0